backend/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.3
pandas
numpy
Django>=5.0,<5.1
djangorestframework
djangorestframework-simplejwt
python-decouple
black
django-cors-headers